Extract tournaments class
