+++ html/pool-tracker/index.html
@@ -80,7 +80,7 @@
         </div>
     </div>
 
-    <script src="app.js"></script>
+    <script src="app.js" type="module"></script>
 </body>
 
 </html>
